refactor(revision): extract revision comparison into a concern
Pure move of compare_types_de_champ, compare_ineligibilite_rules and
their private helpers from ProcedureRevision into
RevisionComparisonConcern. No behavior change.
